api-desafio/core/urls.py
from django.urls import include, path
from django.contrib import admin
from rest_framework import routers
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
from api import views


router = routers.DefaultRouter()
router.register('user', views.UserViewSet)

urlpatterns = [
    path('api/', include(router.urls)),
    path('api/login/', include('rest_framework.urls', namespace='rest_framework')),
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/refresh-token/', TokenRefreshView.as_view(), name='token_refresh')
]